+++ src/api/sql_analysis/index.d.ts
@@ -1,6 +1,14 @@
-import { IGetSQLAnalysisReq, IDirectGetSQLAnalysisResV1 } from '../common.d';
+import { IDirectGetSQLAnalysisResV1 } from '../common.d';
 
-export interface IDirectGetSQLAnalysisV1Params extends IGetSQLAnalysisReq {}
+export interface IDirectGetSQLAnalysisV1Params {
+  project_name: string;
+
+  instance_name: string;
+
+  schema_name?: string;
+
+  sql?: string;
+}
 
 export interface IDirectGetSQLAnalysisV1Return
   extends IDirectGetSQLAnalysisResV1 {}