+++ src/page/GlobalRuleTemplate/RuleTemplateList/column.tsx
@@ -12,6 +12,7 @@ import { t } from '../../../locale';
 import { TableColumn } from '../../../types/common.type';
 import { RuleUrlParamKey } from '../../Rule/useRuleFilterForm';
 import { Link } from '../../../components/Link';
+import DatabaseTypeLogo from '../../../components/DatabaseTypeLogo';
 
 export const RuleTemplateListTableColumnFactory = (
   deleteTemplate: (name: string) => void,
@@ -43,6 +44,13 @@ export const RuleTemplateListTableColumnFactory = (
     {
       dataIndex: 'db_type',
       title: () => t('ruleTemplate.ruleTemplateList.table.dbType'),
+      render(type: string) {
+        if (!type) {
+          return '--';
+        }
+
+        return <DatabaseTypeLogo dbType={type} />;
+      },
     },
     {
       dataIndex: 'operator',